db: Change mechanism for claiming queued runs
A good read on concurrency:
  https://www.percona.com/blog/2014/09/11/openstack-users-shed-light-on-percona-xtradb-cluster-deadlock-issues

Turns out select-for-update isn't fool-proof. Productions instances
show a fair amount of errors like:

 Deadlock found when trying to get lock; try restarting transaction

By doing an update that's limited to one row. The code can check to
see if it was the one that made the change. If so, that worker owns
the Run.

Signed-off-by: Andy Doan <andy.doan@linaro.org>