models: Update Worker.name constraint
While testing MySQL with Galera. Specifically the docker image:

 capttofu/percona_xtradb_cluster_5_6

it was discovered that this version of MySQL won't allow for 1024
primary key length. 1024 was arbitrary and 512 should be fine (mysql
claimed to allow 768).

Given the fact there are no real users of this and we are about to do
a clean deployment, I've chosen to just update models.py and our
original migration by hand to hack around this.

Signed-off-by: Andy Doan <andy.doan@linaro.org>