models: Fix pop_queued ordering
This fixes a weird issue for builds that trigger runs. The issue I saw
was that we had a zephyr-upstream build(1) trigger with runs like

 foo
 bar

After those 2 runs were created a new zephyr build(2) was triggered. At
this point build(1)'s foo run completed and triggered a LAVA test.
Rather than working through this LAVA test, we had build(2)'s runs
execute and then the LAVA test. This makes sure we complete the runs
in build before going to the next one.

Signed-off-by: Andy Doan <andy@opensourcefoundries.com>