email: Fix concurrency issues with email/tests
We are seeing builds that have tests from LAVA often fail to send
out build-complete emails. This appears to be caused by the
lava-reaper sending a bunch of test-completed updates and then the
the JobServ API threads getting confused about when the build was
completed. So ultimately the build's status will be PASSED, but none
of the threads know who completed it and fail to send the email.

Adding a with build.locked(r) to test.py didn't make it safe. To top
that off, the "locked" logic is insane. There's got to be a cleaner
way relying on MySQL, but with our current multi-master approach this
makes things a lot easier.

Signed-off-by: Andy Doan <andy@opensourcefoundries.com>