workers: Make surge-support understand host tags
The current surge-support logic is too dumb and the ratio is too high.
With the current ratio of 4, you can have 20 amd64 runs in QUEUE. If
you have 3 amd64 workers, 1 armhf, and 1 aarch64, the logic will see
the queue ratio 20 / 5 total workers and not enable the surge.

This new approach does per host-tag logic to handle situations like
the one described above. It doesn't yet handle a project-defintion
that uses a wild-card like a%. However, this should help a lot and
we aren't using wild-cards in production.

I'm sure Knuth could find a better way to implement this, but for the
queue sizes and number of workers a CI system sees, I doubt efficiency
matters too much here.

Signed-off-by: Andy Doan <andy@opensourcefoundries.com>