schema: Add a project-wide email setting
The current email schema makes it hard to set an email for triggers
that trigger other runs. eg:

triggers:
  - name: slimmed down example
    runs:
      - name: compile-{loop}
        loop-on:
          - param: PLATFORM
            values: [frdm_k64f, 96b_nitrogen]
        triggers:
          - name: lava-test
            run-names: "{name}-{loop}"
    email:
      users: 'foo@example.com'

  - name: lava-test
    type: lava
    runs:
      - name: lava

The intent of this author is probably to notify at the end of the
lava triggered runs. However, the author would have to add an email
clause to the lava-test and then update the current email clause to
only notify on failures to get the effect of what a project-wide email
setting would do.

I considered removing the other email stanzas (per run and trigger),
but they *could* be handy and have unit-test code, so I've left them.

Signed-off-by: Andy Doan <andy@opensourcefoundries.com>