runner: Always create a netrc file / include .curlrc
Since the JobServ now has the ability to force authentication on GET
requests, the runner should include the run's api token for requests
against the JobServ. Populating this in the .netrc makes this easier
for almost all users and including a .curlrc makes it transparent
for anyone using curl inside their runs.

Signed-off-by: Andy Doan <andy@opensourcefoundries.com>