worker: Create a way to log events from a worker
Workers may have things like cron jobs set up and it would be nice to
give them the ability to log events for success/failure. This call
is very generic and simply logs the message. It could be monkey patched
at runtime to do something more interesting like sending emails or
sending the event to a monitoring system.

Signed-off-by: Andy Doan <andy@opensourcefoundries.com>