runner: Add reboot support
This adds a simple mechanism that allows a run to cause the server
to do a reboot and resume. If a run creates a file called
/archive/execute-on-reboot, the handler and jobserv-worker will set
things up reboot and resume the run.

Since this could easily be (ab)used, the jobserv-worker script does
not call "sudo reboot". ie - A non-privileged worker (not running as
root) won't be able to reboot and the run will fail.

Signed-off-by: Andy Doan <andy@opensourcefoundries.com>