runner: Add "cold" reboot support
This change allows a worker to handle a run that requests a cold
reboot. This requires that the worker itself provide a script/binary
named /usr/bin/cold-reboot that can interface with a PDU in a way
that can power can be turned off and then back on for the worker.

Signed-off-by: Andy Doan <andy@opensourcefoundries.com>