Move "internal_api" logic into permissions module
The internal api stuff was a hack created before the JobServ had
a pluggable concept for dealing with authentication/authorization. Now
that we have a permissions module, it makes sense to make the internal
api stuff something that can be controlled in someone's deployment
using their own security measures.

This moves the first half of the internal_requests out.

Signed-off-by: Andy Doan <andy@opensourcefoundries.com>