models: Add a better accessor for trigger secrets
This was bad from the begging, but secrets in triggers were originally
going to get replaced with something else. We've kept them, so lets
start cleaning up how we use them.

Once code is refactored to use these new methods, we'll be able to
encrypt the data we store in the DB

Signed-off-by: Andy Doan <andy@opensourcefoundries.com>