worker: Give rebooted runs a unique name
In theory this isn't needed. Its really for debug help. We have reboot
runs where runs/rebooted-run exists when the device comes up and the
next rebooted run under /srv/jobserv-worker can't be renamed and
executed. This should help cure that. It will also leave around the
old rebooted runs so we have some history of what runs failed to
complete.

In other words - this is a diagnostic commit.

Signed-off-by: Andy Doan <andy@opensourcefoundries.com>