git-poller: Handle "not found" case correctly
The current logic made the poller think *any* error meant that we
assumed it was the first time the poller had been. This caused us
to clear out the cache we had and start over. This change properly
detects when the cache isn't found and handles it w/o causing us
to overwrite our cache.

Signed-off-by: Andy Doan <andy@foundries.io>