worker: Add in optional non-forking "looper"
The original approach with the worker had two things in mind:

 * Avoiding thinking about daemon logic
 * Avoiding potential daemon issues like memory leaks

The problem with this is that forking a call to "check" every X seconds
is a little costly and makes systems under test a little jittery. The
python runtime is also a little slow to load on something as small
as an rpi0-wifi which we run Workers on. Additionally, its just too
easy to use SystemD and let it handle daemon logic for you. Lastly,
this is so small and in Python3 - its impossible to have a leak :)

This adds in the ability via a feature flag so I can start testing
this in the wild to make sure it works well before making it the
default way.

Signed-off-by: Andy Doan <andy@foundries.io>