run: Allow triggers to be "upgraded"
We could have a build that's triggered by either a github_pr or a
gitlab_mr. They might have runs that trigger something of type
"simple". This could be the case where a git_poller and github_mr both
trigger a similar set of tests *after* a build. In the case of the
github_pr, we should "upgrade" the type of each run from simple to
github_pr so that it can update the status of the PR.

Signed-off-by: Andy Doan <andy@foundries.io>