migrations: Fix sqlite3 issue for deleted worker
sqlite does not allow for the non-null constraint when altering a
table. As it turns out, it was probably dumb to make this requirement
in the first place. "null" is just going to evaluate like "false" for
our worker.deleted logic, and all the other boolean columns we've
added as migrations like surges_only were done the same way.

So this won't affect anyone with an existing deployment, but will
make things easier for people toying with sqlite

Signed-off-by: Andy Doan <andy@foundries.io>