worker: Clean up broken locks logic
We are seeing that a worker running a job will hold all the flocks,
not just the one it popped. I think this is probably due to the fact
we are forking and the reference to the other fd's are getting held
by accident.

This change adds a more fool proof way of dealing with these locks
so that they are explicitly closed each time. It also writes a
value to the file just to ease debug work.

Signed-off-by: Andy Doan <andy@foundries.io>