github: Handle multiple triggers
We never considered the case where a single JobServ Project might
have more than one GitHub Triggers assigned to it. This has now
become a reality with the introduction of the "lmp" project having
a trigger for containers.git and lmp-manifest.git.

This change loops through all the Triggers hoping to find the correct
one.

Signed-off-by: Andy Doan <andy@foundries.io>