models: Handle synchronous project builds better
There's a problem with the current logic: A synchronous project with
a build containing multiple runs (say multi-arch docker containers)
will run sequentially.

This query is becoming insane and I think the new approach is probably
going to be *more* efficient since its so much less stress on the
database coupled with the fact that we are never going to have that
many concurrent runs.

The new approach is to find out everything we need about QUEUED and
RUNNING runs. We can then loop through these results in a single
pass to know if there's a run we can schedule.

Signed-off-by: Andy Doan <andy@foundries.io>