Fix synchronous project build logic with tags
We hit the following issue in production with a synchronous project:

* long running build with one run for the amd64 host-tag is RUNNING
* a new build for this project is QUEUED with:
  run-1 - host-tag=amd64
  run-2 - host-tag=armhf
  run-3 - host-tag=aarch64

The query we have to find candidates is filtering on host-tag, so we
accidentally allowed run-2 and run-3 to get executed before the
first long running build was done.

The code did handle the "run-1" properly since their host tags matched.

The fix here is to handle the host-tag filtering *after* we find our
candidates. Since we no longer filter inside SQL, we can go back to
letting our host-tags be in python's fnmatch style format.

I also renamed the runs in the test to make them a bit
easier to understand.

Signed-off-by: Andy Doan <andy@foundries.io>