runner: Improve command polling logic
pyton3.4 and greater return an empty list, so we can be simplify how we
call the poll api.

Additionally the timeout value was wrong. It was assuming seconds and
the expected value is milliseconds. This leads to some aggressive
polling and we really don't have anything useful to do while we wait on
output.

Signed-off-by: Andy Doan <andy@foundries.io>