handlers: Try https on github submodules
We often have users with git submodules that try to pull down
repositories like:

 git@github.com:<company>/repo

This is great for local development but doesn't work well in CI. In the
past we've suggested users set their submodules to https. This is a bit
of a pain for them, and this change allows us to manage it.

Signed-off-by: Andy Doan <andy@foundries.io>