runner: Use real .gitconfig
The complexity of our git configuration is growing. We've already
started (ab)using $HOME for .netrc. This does the same to build up a
.gitconfig that will be easier to understand.

This also fixes a real issue. The "-c http.extraheaders" breaks
repositories with git-lfs. git-lfs was only working with submodules.
This feels like a bug in git-lfs but this fixes it also.

Signed-off-by: Andy Doan <andy@foundries.io>